requirements.txt
# Core dependencies for HAILO AI HAT development
# PyHailoRT - Python bindings for HailoRT
hailo-platform>=4.17.0

# Computer Vision and AI
opencv-python>=4.8.0
numpy>=1.21.0
pillow>=9.0.0

# Data processing
pandas>=1.5.0

# Configuration and utilities
pyyaml>=6.0
click>=8.0.0
python-dotenv>=0.19.0

# Logging and monitoring
loguru>=0.6.0

# Testing (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0

# Development tools (optional)
black>=22.0.0
flake8>=5.0.0
mypy>=0.991
